+++ src/create-commands.ts
@@ -2,6 +2,7 @@
 import * as vscode from 'vscode';
 import * as chokidar from 'chokidar';
 import * as commands from './commands.json';
+import * as path from 'path';
 
 export default commands.map(({ title, command, extension }) => ({ title, extension, command: () => selectParentFolderAndRun(command) }));
 
@@ -10,15 +11,37 @@ async function selectParentFolderAndRun(command: string) {
     if (!projectFolder) {
         return;
     }
-    const [parentPath, projectName] = projectFolder;
+    const [parentPath, projectName, uri] = projectFolder;
+    const result = await selectVSCodeWindow(uri);
     await executeCreateCommand(parentPath, projectName, command);
+}
+
+async function selectVSCodeWindow(path: vscode.Uri) {
+    // TODO: actually make the new window flow work!
+    vscode.workspace.updateWorkspaceFolders(0, 0, {
+        uri: path,
+    });
+
+    // const option = await vscode.window.showQuickPick([
+    //     'Open in current workspace',
+    //     'Open in new window',
+    //     'Cancel'],
+    // );
+
+    // if (option === 'Open in current workspace') {
+    //     vscode.workspace.updateWorkspaceFolders(0, 0, {
+    //         uri: path,
+    //     });
+    // } else if (option === 'Open in new window') {
+    //     await vscode.commands.executeCommand('vscode.openFolder', path);
+    // }
 }
 
 // chokidar.watch('.', {depth: 0, ignoreInitial: true }).on('addDir', function(path) {console.log('File', path, 'has been added');});
 
-async function selectProjectFolder(): Promise<string[] | null> {
-    const value = await vscode.window.showQuickPick(["Select Project Folder", "Cancel"], {
-        placeHolder: "Select Project Folder",
+async function selectProjectFolder(): Promise<[string, string, vscode.Uri] | null> {
+    const value = await vscode.window.showQuickPick(["Select project folder", "Cancel"], {
+        placeHolder: "Select project folder",
     });
     if (value === "Cancel") {
         return null;
@@ -33,10 +56,11 @@ async function selectProjectFolder(): Promise<string[] | null> {
         return null;
     }
     const pathToProject = result[0].path;
-    const projectName = pathToProject.split("/").slice(-1)[0];
-    const parentPath = pathToProject.substring(0, pathToProject.lastIndexOf('/'));
+    const seperator = path.sep;
+    const projectName = pathToProject.split(seperator).slice(-1)[0];
+    const parentPath = pathToProject.substring(0, pathToProject.lastIndexOf(seperator));
 
-    return [parentPath, projectName];
+    return [parentPath, projectName, result[0]];
 }
 
 async function executeCreateCommand(path: string, projectName: string, command: string) {