+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
 ### 0.0.1
+- (Bugfix) Fixed seperator for windows
+- (UX) Opening new project in current workspace
+
+### 0.0.1
 
 Init release with starters for:
 - Django